// discovery/plugins/notifications/src/types/notifications.ts
export type DMNotification = {
  chat_id: string
  sender_user_id: number
  receiver_user_id: number
  timestamp: Date
}

export type DMReactionNotification = {
  chat_id: string
  message_id: string
  sender_user_id: number
  receiver_user_id: number
  reaction: string
  timestamp: Date
}